Move PostmasterProcess to a separate file
